Refactor CLI commands to delegate to reusable operations
